requirements.txt
azure-identity
azure-mgmt-compute
azure-mgmt-monitor
requests
